Refactoring with Extract class by extracting the CSVEscape implementation and using it seperately outside of StringEscapeUtils
